Upgrade to tsx
