+++ package.json
@@ -13,10 +13,10 @@
   },
   "scripts": {
     "build": "vite build",
-    "release": "npm run build && npm run build:ts && npm run build:dts && npm run build:demo",
-    "build:demo": "vite build -c vite.demo.config.js",
+    "release": "npm run build && npm run build:resolver && npm run build:ts && npm run build:dts",
+    "build:resolver": "vite build -c vite.resolver.config.js",
     "build:ts": "NODE_ENV=production tsc --emitDeclarationOnly -p ./tsconfig.build.json",
-    "build:dts": "api-extractor run --local --verbose && tail -n +3 src/globalExtensions.ts >> dist/index.d.ts",
+    "build:dts": "api-extractor run --local --verbose && tail -n +3 src/global-extensions.ts >> dist/index.d.ts && mv dist/src/auto-import-resolver.d.ts dist/auto-import-resolver.d.ts",
     "dev": "vite serve -c vite.demo.config.js",
     "unit": "vitest",
     "unit:watch": "vitest  --watch",
@@ -64,7 +64,8 @@
     "@microsoft/api-extractor": "^7.33.5",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.41.0",
-    "@vitejs/plugin-vue": "^4.2.3",
+    "@vitejs/plugin-vue": "^4.5.2",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
     "@vue/compiler-sfc": "^3.3.4",
     "@vue/server-renderer": "^3.2.37",
     "@vue/test-utils": "^2.2.1",
@@ -75,10 +76,11 @@
     "mitt": "^3.0.0",
     "sass": "^1.57.1",
     "typescript": "^5.0.4",
+    "unplugin-vue-components": "^0.26.0",
     "velocity-animate": "^1.5.2",
-    "vite": "^4.3.8",
-    "vite-plugin-css-injected-by-js": "^3.1.1",
-    "vitest": "^0.31.1",
+    "vite": "^5.0.6",
+    "vite-plugin-css-injected-by-js": "^3.3.0",
+    "vitest": "^1.0.2",
     "vue": "^3.3.4",
     "vuepress": "^2.0.0-beta.67"
   }
