+++ src/components/Notifications.css
@@ -1,0 +1,60 @@
+.vue-notification-group {
+  display: block;
+  position: fixed;
+  z-index: 5000;
+}
+  
+.vue-notification-wrapper {
+  display: block;
+  overflow: hidden;
+  width: 100%;
+  margin: 0;
+  padding: 0;
+}
+  
+.notification-title {
+  font-weight: 600;
+}
+  
+.vue-notification-template {
+  display: block;
+  box-sizing: border-box;
+  background: white;
+  text-align: left;
+}
+  
+.vue-notification {
+  display: block;
+  box-sizing: border-box;
+  text-align: left;
+  font-size: 12px;
+  padding: 10px;
+  margin: 0 5px 5px;
+  
+  color: white;
+  background: #44A4FC;
+  border-left: 5px solid #187FE7;
+}
+  
+.vue-notification.warn {
+  background: #ffb648;
+  border-left-color: #f48a06;
+}
+  
+.vue-notification.error {
+  background: #E54D42;
+  border-left-color: #B82E24;
+}
+  
+.vue-notification.success {
+  background: #68CD86;
+  border-left-color: #42A85F;
+}
+  
+.vn-fade-enter-active, .vn-fade-leave-active, .vn-fade-move  {
+  transition: all .5s;
+}
+  
+.vn-fade-enter-from, .vn-fade-leave-to {
+  opacity: 0;
+}
