--- src/components/group/CssGroup.vue
@@ -1,18 +1,0 @@
-<template>
-  <transition-group
-    tag="span"
-    :name="name"
-  >
-    <slot />
-  </transition-group>
-</template>
-
-<script setup lang="ts">
-defineOptions({
-  inheritAttrs: false,
-});
-
-defineProps<{
-  name: string
-}>();
-</script>
