+++ src/plugin.ts
@@ -1,14 +1,15 @@
 
-import { App } from 'vue';
-import Notifications from '@/components/Notifications.vue';
+import type { App } from 'vue';
+import Notifications from '@/components/Notifications';
 import { params } from '@/params';
-import { NotificationsPluginOptions } from '@/types';
+import type { NotificationsPluginOptions } from '@/types';
 import { notify } from '@/notify';
+import { COMPONENT_NAME } from '@/constants';
 
 export function install(app: App, args: NotificationsPluginOptions = {}): void {
   Object.entries(args).forEach((entry) => params.set(...entry));
   const name = args.name || 'notify';
 
   app.config.globalProperties['$' + name] = notify;
-  app.component(args.componentName || 'Notifications', Notifications);
+  app.component(args.componentName || COMPONENT_NAME, Notifications);
 }
